Structured demo workbook into same structure as file system Removed commented out code Ensured all demos work with latest SDK Removed framework from Example folder, and added README.md instructing users to download the latest release themselves.
