+++ Examples/Authentication/SafariViewControllerDemo/SafariViewControllerDemo/Info.plist
@@ -25,8 +25,6 @@
 			<string></string>
 			<key>CFBundleURLSchemes</key>
 			<array>
-			<key>CFBundleURLSchemes</key>
-			<array>
 				<string>predixsdk</string>
 			</array>
 		</dict>
