--- Examples/Frameworks/PredixSDK.framework/Headers/Extensions.h
@@ -1,8 +1,0 @@
-//
-// Created by Johns, Andy (GE Corporate) on 3/26/15.
-// Copyright (c) 2015 General Electric Company. All rights reserved.
-//
-
-@interface NSString(SHA512)
-- (NSString * _Nonnull)getSHA512Hash;
-@end
