--- Examples/Frameworks/PredixSDK.framework/Headers/PredixSDK-Swift.h
@@ -1,295 +1,0 @@
-// Generated by Apple Swift version 4.0.3 (swiftlang-900.0.74.1 clang-900.0.39.2)
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wgcc-compat"
-
-#if !defined(__has_include)
-# define __has_include(x) 0
-#endif
-#if !defined(__has_attribute)
-# define __has_attribute(x) 0
-#endif
-#if !defined(__has_feature)
-# define __has_feature(x) 0
-#endif
-#if !defined(__has_warning)
-# define __has_warning(x) 0
-#endif
-
-#if __has_attribute(external_source_symbol)
-# define SWIFT_STRINGIFY(str) #str
-# define SWIFT_MODULE_NAMESPACE_PUSH(module_name) _Pragma(SWIFT_STRINGIFY(clang attribute push(__attribute__((external_source_symbol(language="Swift", defined_in=module_name, generated_declaration))), apply_to=any(function, enum, objc_interface, objc_category, objc_protocol))))
-# define SWIFT_MODULE_NAMESPACE_POP _Pragma("clang attribute pop")
-#else
-# define SWIFT_MODULE_NAMESPACE_PUSH(module_name)
-# define SWIFT_MODULE_NAMESPACE_POP
-#endif
-
-#if __has_include(<swift/objc-prologue.h>)
-# include <swift/objc-prologue.h>
-#endif
-
-#pragma clang diagnostic ignored "-Wauto-import"
-#include <objc/NSObject.h>
-#include <stdint.h>
-#include <stddef.h>
-#include <stdbool.h>
-
-#if !defined(SWIFT_TYPEDEFS)
-# define SWIFT_TYPEDEFS 1
-# if __has_include(<uchar.h>)
-#  include <uchar.h>
-# elif !defined(__cplusplus) || __cplusplus < 201103L
-typedef uint_least16_t char16_t;
-typedef uint_least32_t char32_t;
-# endif
-typedef float swift_float2  __attribute__((__ext_vector_type__(2)));
-typedef float swift_float3  __attribute__((__ext_vector_type__(3)));
-typedef float swift_float4  __attribute__((__ext_vector_type__(4)));
-typedef double swift_double2  __attribute__((__ext_vector_type__(2)));
-typedef double swift_double3  __attribute__((__ext_vector_type__(3)));
-typedef double swift_double4  __attribute__((__ext_vector_type__(4)));
-typedef int swift_int2  __attribute__((__ext_vector_type__(2)));
-typedef int swift_int3  __attribute__((__ext_vector_type__(3)));
-typedef int swift_int4  __attribute__((__ext_vector_type__(4)));
-typedef unsigned int swift_uint2  __attribute__((__ext_vector_type__(2)));
-typedef unsigned int swift_uint3  __attribute__((__ext_vector_type__(3)));
-typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
-#endif
-
-#if !defined(SWIFT_PASTE)
-# define SWIFT_PASTE_HELPER(x, y) x##y
-# define SWIFT_PASTE(x, y) SWIFT_PASTE_HELPER(x, y)
-#endif
-#if !defined(SWIFT_METATYPE)
-# define SWIFT_METATYPE(X) Class
-#endif
-#if !defined(SWIFT_CLASS_PROPERTY)
-# if __has_feature(objc_class_property)
-#  define SWIFT_CLASS_PROPERTY(...) __VA_ARGS__
-# else
-#  define SWIFT_CLASS_PROPERTY(...)
-# endif
-#endif
-
-#if __has_attribute(objc_runtime_name)
-# define SWIFT_RUNTIME_NAME(X) __attribute__((objc_runtime_name(X)))
-#else
-# define SWIFT_RUNTIME_NAME(X)
-#endif
-#if __has_attribute(swift_name)
-# define SWIFT_COMPILE_NAME(X) __attribute__((swift_name(X)))
-#else
-# define SWIFT_COMPILE_NAME(X)
-#endif
-#if __has_attribute(objc_method_family)
-# define SWIFT_METHOD_FAMILY(X) __attribute__((objc_method_family(X)))
-#else
-# define SWIFT_METHOD_FAMILY(X)
-#endif
-#if __has_attribute(noescape)
-# define SWIFT_NOESCAPE __attribute__((noescape))
-#else
-# define SWIFT_NOESCAPE
-#endif
-#if __has_attribute(warn_unused_result)
-# define SWIFT_WARN_UNUSED_RESULT __attribute__((warn_unused_result))
-#else
-# define SWIFT_WARN_UNUSED_RESULT
-#endif
-#if __has_attribute(noreturn)
-# define SWIFT_NORETURN __attribute__((noreturn))
-#else
-# define SWIFT_NORETURN
-#endif
-#if !defined(SWIFT_CLASS_EXTRA)
-# define SWIFT_CLASS_EXTRA
-#endif
-#if !defined(SWIFT_PROTOCOL_EXTRA)
-# define SWIFT_PROTOCOL_EXTRA
-#endif
-#if !defined(SWIFT_ENUM_EXTRA)
-# define SWIFT_ENUM_EXTRA
-#endif
-#if !defined(SWIFT_CLASS)
-# if __has_attribute(objc_subclassing_restricted)
-#  define SWIFT_CLASS(SWIFT_NAME) SWIFT_RUNTIME_NAME(SWIFT_NAME) __attribute__((objc_subclassing_restricted)) SWIFT_CLASS_EXTRA
-#  define SWIFT_CLASS_NAMED(SWIFT_NAME) __attribute__((objc_subclassing_restricted)) SWIFT_COMPILE_NAME(SWIFT_NAME) SWIFT_CLASS_EXTRA
-# else
-#  define SWIFT_CLASS(SWIFT_NAME) SWIFT_RUNTIME_NAME(SWIFT_NAME) SWIFT_CLASS_EXTRA
-#  define SWIFT_CLASS_NAMED(SWIFT_NAME) SWIFT_COMPILE_NAME(SWIFT_NAME) SWIFT_CLASS_EXTRA
-# endif
-#endif
-
-#if !defined(SWIFT_PROTOCOL)
-# define SWIFT_PROTOCOL(SWIFT_NAME) SWIFT_RUNTIME_NAME(SWIFT_NAME) SWIFT_PROTOCOL_EXTRA
-# define SWIFT_PROTOCOL_NAMED(SWIFT_NAME) SWIFT_COMPILE_NAME(SWIFT_NAME) SWIFT_PROTOCOL_EXTRA
-#endif
-
-#if !defined(SWIFT_EXTENSION)
-# define SWIFT_EXTENSION(M) SWIFT_PASTE(M##_Swift_, __LINE__)
-#endif
-
-#if !defined(OBJC_DESIGNATED_INITIALIZER)
-# if __has_attribute(objc_designated_initializer)
-#  define OBJC_DESIGNATED_INITIALIZER __attribute__((objc_designated_initializer))
-# else
-#  define OBJC_DESIGNATED_INITIALIZER
-# endif
-#endif
-#if !defined(SWIFT_ENUM_ATTR)
-# if defined(__has_attribute) && __has_attribute(enum_extensibility)
-#  define SWIFT_ENUM_ATTR __attribute__((enum_extensibility(open)))
-# else
-#  define SWIFT_ENUM_ATTR
-# endif
-#endif
-#if !defined(SWIFT_ENUM)
-# define SWIFT_ENUM(_type, _name) enum _name : _type _name; enum SWIFT_ENUM_ATTR SWIFT_ENUM_EXTRA _name : _type
-# if __has_feature(generalized_swift_name)
-#  define SWIFT_ENUM_NAMED(_type, _name, SWIFT_NAME) enum _name : _type _name SWIFT_COMPILE_NAME(SWIFT_NAME); enum SWIFT_COMPILE_NAME(SWIFT_NAME) SWIFT_ENUM_ATTR SWIFT_ENUM_EXTRA _name : _type
-# else
-#  define SWIFT_ENUM_NAMED(_type, _name, SWIFT_NAME) SWIFT_ENUM(_type, _name)
-# endif
-#endif
-#if !defined(SWIFT_UNAVAILABLE)
-# define SWIFT_UNAVAILABLE __attribute__((unavailable))
-#endif
-#if !defined(SWIFT_UNAVAILABLE_MSG)
-# define SWIFT_UNAVAILABLE_MSG(msg) __attribute__((unavailable(msg)))
-#endif
-#if !defined(SWIFT_AVAILABILITY)
-# define SWIFT_AVAILABILITY(plat, ...) __attribute__((availability(plat, __VA_ARGS__)))
-#endif
-#if !defined(SWIFT_DEPRECATED)
-# define SWIFT_DEPRECATED __attribute__((deprecated))
-#endif
-#if !defined(SWIFT_DEPRECATED_MSG)
-# define SWIFT_DEPRECATED_MSG(...) __attribute__((deprecated(__VA_ARGS__)))
-#endif
-#if __has_feature(attribute_diagnose_if_objc)
-# define SWIFT_DEPRECATED_OBJC(Msg) __attribute__((diagnose_if(1, Msg, "warning")))
-#else
-# define SWIFT_DEPRECATED_OBJC(Msg) SWIFT_DEPRECATED_MSG(Msg)
-#endif
-#if __has_feature(modules)
-@import ObjectiveC;
-@import Foundation;
-#endif
-
-#pragma clang diagnostic ignored "-Wproperty-attribute-mismatch"
-#pragma clang diagnostic ignored "-Wduplicate-method-arg"
-#if __has_warning("-Wpragma-clang-attribute")
-# pragma clang diagnostic ignored "-Wpragma-clang-attribute"
-#endif
-#pragma clang diagnostic ignored "-Wunknown-pragmas"
-#pragma clang diagnostic ignored "-Wnullability"
-
-SWIFT_MODULE_NAMESPACE_PUSH("PredixSDK")
-
-/// Provides Logging support for the SDK using the Apple Unified Logging API
-/// seealso:
-///
-/// <ul>
-///   <li>
-///     <a href="https://developer.apple.com/reference/os/1891852-logging">Logging - os | Apple Developer Documentation</a>
-///   </li>
-///   <li>
-///     <a href="https://developer.apple.com/videos/play/wwdc2016/721/">WWDC 2016: Unified Logging and Activity Tracing</a>
-///   </li>
-/// </ul>
-SWIFT_CLASS("_TtC9PredixSDK6Logger")
-@interface Logger : NSObject
-/// Initializes a new logger object
-/// This Logger will automatically attempt to load it’s inital logging level setting from configuration according to
-/// the <code>Logger.loggingLevelConfigLocation</code> and <code>Logger.loggingLevelConfigKey</code> values.
-/// If no configured value is found, the <code>Logger.defaultLoggingLevel</code> setting will be used.
-/// If the configuration location is Settings, an observer will be created to monitor for changes to the
-/// current value and automatically update the loggingLevel as Settings changes.
-- (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
-@end
-
-
-
-
-/// Allows an application developer to provide mock data responses for URLSession and other types of network request classes (UIWebView for example).
-/// The MockNetworkDataManager works with mock data providers to determine if a requested URL should return real network data or mock network data.
-/// By default the MockNetworkDataManager uses a default data provider that works with a resource bundle and plist to determine if a requested URL should return mock data.
-/// Here is an example of how the default provider works:
-/// In a project define a MockData.bundle resource bundle, inside that bundle it should contain all of the contents you need to mock a network response for a request.  The bundle should also include a
-/// DataMapping.plist file.
-/// <em>Example:</em>
-/// <?xml version="1.0" encoding="UTF-8"?>
-/// <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-/// <plist version="1.0">
-/// <dict>
-///     <key>https://apple.com</key>
-///     <dict>
-///         <key>dataFilename</key>
-///         <string>mockData.json</string>
-///         <key>requests</key>
-///         <array>
-///             <dict>
-///                 <key>httpMethod</key>
-///                 <string>GET</string>
-///                 <key>headerFields</key>
-///                 <dict>
-///                     <key>a</key>
-///                     <string>b</string>
-///                 </dict>
-///                 <key>response</key>
-///                 <dict>
-///                     <key>headerFields</key>
-///                     <dict>
-///                         <key>a</key>
-///                         <string>b</string>
-///                     </dict>
-///                     <key>httpVersion</key>
-///                     <string>1.1</string>
-///                     <key>statusCode</key>
-///                     <integer>200</integer>
-///                 </dict>
-///             </dict>
-///             <dict>
-///                 <key>httpMethod</key>
-///                 <string>POST</string>
-///             </dict>
-///         </array>
-///     </dict>
-/// </plist>
-/// Alternatively you can define a simple response structure, this will return the response for ANY requests sent for the given endpoint
-/// <?xml version="1.0" encoding="UTF-8"?>
-/// <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-/// <plist version="1.0">
-/// <dict>
-///     <key>https://simple.com</key>
-///     <dict>
-///         <key>dataFilename</key>
-///         <string>mockData.json</string>
-///         <key>response</key>
-///         <dict>
-///             <key>headerFields</key>
-///             <dict>
-///                 <key>a</key>
-///                 <string>b</string>
-///             </dict>
-///             <key>httpVersion</key>
-///             <string>1.1</string>
-///             <key>statusCode</key>
-///             <integer>200</integer>
-///         </dict>
-///     </dict>
-/// </dict>
-/// </plist>
-/// The data is loaded from disk as raw Data using the contents of file as the data response.  This means the default provider should be able to work with most
-/// file types (JSON, text, images, video, etc.)
-SWIFT_CLASS("_TtC9PredixSDK22MockNetworkDataManager")
-@interface MockNetworkDataManager : NSObject
-- (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
-@end
-
-
-
-
-
-SWIFT_MODULE_NAMESPACE_POP
-#pragma clang diagnostic pop
