--- Examples/Frameworks/PredixSDK.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module PredixSDK {
-  umbrella header "PredixSDK.h"
-
-  export *
-  module * { export * }
-}
-
-module PredixSDK.Swift {
-    header "PredixSDK-Swift.h"
-    requires objc
-}
