+++ Examples/GettingStarted/GettingStarted/ViewController.swift
@@ -13,6 +13,4 @@ class ViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
 
-
 }
-
