--- Examples/OnlineAPI/OnlineAPIDemo/AppDelegate.swift
@@ -1,7 +1,0 @@
-import UIKit
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    var window: UIWindow?
-}
-
