--- Examples/OnlineAPI/OnlineAPIDemo/CouchbaseLite/module.modulemap
@@ -1,3 +1,0 @@
-module CouchbaseLite {
-    export *
-}
