+++ Examples/OnlineAPI/OnlineAPIDemo/OnlineAPIDemo/CouchbaseLite/module.modulemap
@@ -1,0 +1,3 @@
+module CouchbaseLite {
+    export *
+}
