+++ Examples/README.md
@@ -1,0 +1,15 @@
+# PredixSDKforiOS Example projects
+
+**Important** 
+
+Before using these examples you need to download the latest version of the PredixSDKForiOS framework from [Github releases](https://github.com/PredixDev/PredixSDKForiOS/releases).
+
+Unzip the released framework and copy it to the [Frameworks](Frameworks) folder in the [Examples](.) directory.
+
+###Using the Examples
+
+This folder contains many sample projects, you can open indiviual projects in Xcode, or you can open all the examples using the workspace file, [PredixSDKForiOSExamples](PredixSDKForiOSExamples.xcworkspace)
+
+The examples are grouped into folders for the functionality being demonstrated, each functionality folder may have multple example projects within it. 
+
+Additional README.md files within the example folders can provide additional information, where appropriate.
